refactor(herminvestcli): extract add, delete, update and query commands from hermInvestCli
